added admin API routes
