+++ app/api/admin/sending-groups/route.ts
@@ -1,0 +1,4 @@
+import { NextResponse } from "next/server";
+import db from "@/db";
+
+export async function GET(req: Request) {}
