+++ tsconfig.json
@@ -22,6 +22,15 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/api/sending-groups/create/route.js", "app/api/sending-groups/[id]/route.js", "app/api/sending-groups/route.js", "app/layout.jsx"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "app/api/sending-groups/create/route.js",
+    "app/api/sending-groups/[id]/route.js",
+    "app/api/sending-groups/route.js",
+    "app/layout.jsx"
+  ],
   "exclude": ["node_modules"]
 }
